requirements.txt
flask==2.3.3
pandas==1.5.3
numpy==1.24.3
requests==2.31.0
plotly==5.15.0
gunicorn==21.2.0
python-dotenv==1.0.0
pytest==7.3.1